xz_utils-5.0.1.recipe: Turn into an actual recipe
I tried a bit, but there's no easy way of building with gcc 2, as the
code makes heavy use of C99 features. Should we really need to build
it with gcc 2, we should at least deal with the local variable
declarations not at the beginning of a block in an automated fashion.