Update man 1.6g patch
Mostly fixes for using the correct paths for package management, like
not using absolute paths for the tools that man invokes, and using
the correct man search paths.

Displayed man pages contain uninterpreted escape sequences. It works
correctly when adding the '-r' option to DEFAULTLESSOPT (or piping to
"less -irs"), but the comment for this variable explicitly notes that
having to do that hints toward a broken setup. So I didn't do that and
we'll see later whether the current Haiku does better.