openjdk: WIP recipe
Doesn't fully build yet. The build runs for a while, but eventually,
when building the jaxws/jaf sources, javac hangs. It blocks trying to
write to stderr, a pipe whose other end isn't read from (by Ant). I
haven't determined yet, whether the hang is an Ant bug and the issue
is javac simply encountering too many issues, or whether javac writing
a lot to stderr is just fine (verbose diagnostics?) and the hang is
actually a Haiku (pipe related?) bug, a combination of the two, or
something else entirely.