im: allow hybrid builds and other cleanups
 * Fix tecmake to set a soname for the libraries

This currenly fails strict-policy checks because the package doesn't
depend on libstdc++.so, but this should be in haiku$secondaryArchSuffix.
I also had to explicitly depend on the proper version of as, otherwise
the one for gcc2 would be picked up. I suspect one of the dependencies
is wrong, but attempts to use haikuporter --why have failed.