Convert bzr to recipe
Some of the work was done as part of issue #30
I did a moderate amount of fixing to the recipe but bzr doesn't
behave very well. So leaving it marked as broken.

It has trouble finding our ssl certs directory.  Our wget also
had trouble with certs using the original url in that issue. This
can be worked around at runtime by setting command line or config
options.

When doing a test clone of the bzr repo it failed to completely
check out the source with some odd out of memory error.  I seem
to recall bzr having this sort of trouble before on Haiku.