Fixed ClipUp recipe.
Improved SUMMARY, corrected SRC_URI, added USER_SETTINGS_FILES,
fixed BUILD() and INSTALL().
Use the "CTRLV" input filter version, as that seems to be the more
natural key combo.

The app still has some focus issues and it doesn't visually confirm
to the rest of the Haiku GUI, but it's basically working.