Add dcron.
There doesn't seem to be a way to autostart the daemon on boot (unless
we directly modify the boot script). Assume the user will do that
himself by adding to ~/config/settings/boot/.

crond needs to be told where to store the crontab and cronstamps file,
and where to read the cron files. The default values are non-existing
folders in /var and /etc (just creating those is ok, too).