autoconf: Break cyclic dependency
There is no reason why autoconf would need autoconf and thus have a
cyclic dependency, as there is no need to call autoreconf as no files
are changed that result in generated files.

It seems the BUILD_PREREQUIRES only included autoconf as a hack to
include perl, awk, m4 and others needed to build autoreconf.