llvm: fix recipe.
Set the installation paths by forcing the variable passed to make; as configure options are ignored.
As a result; add a bunch of provides to match the now properly detected libraries which are installed in alace where people can use them.

This makes the package build properly with haikuporter -S.