ffmpeg 0.11: make it work with gcc2.
* Assembler code completely disabled for gcc2. Cross platform C code is
used instead, which means this needs more CPU than the 0.10.x branch
with asm enabled. But, at least it works.
* Redone the gcc2 support to make it more well behaved. Things that only
* need to be disabled for gcc2 are under a #if __GNUC__ > 2 check,
instead of #if 0. Also removed a lot of hacks to the configure script
that are not needed anymore.
* Since the patch for gcc2 should not create problems for the gcc4
version, use a single patch. The hack used to decide which patchset to
use did not work, because effectoiveTargetarchitecture was not set yet
when parsing the file to find the PATCHES variable.
* I leave this marked as broken for gcc2, because 0.11.5 was not updated
for years, while the ffmpeg team is still updating the 0.10.x branch
with regular releases. This means 0.10 gets bugfixes in the codecs, and
0.11 doesn't. We better use the latest codecs, even if we don't get the
latest API.