Duplicate Python license file in ports requiring it.
While Haiku usually brings Python with it so its license is available
in the merged licenses directory, this isn't declared as a dependency
anywhere and therefore "chance". Since we yet lack a way to declare this
dependency (for lack of a license entity), duplicate the Python license
in the ports requiring it to make them self-contained.