Split flare's data package into its own recipe.
I learned that by PulkoMandy's comment on the GCI task to port openarena:
"With what you did, there will be one data package for each architecture
we compile the game for. This means a waste of space on our servers to
store the data several times.
The data should be a separate recipe as it is done for the openttd game
(see the openttd_sfx recipe for example)."