lighttpd: 1.4.35→1.4.39, ARCHITECTURES += ?x86 ?x86_64 ?arm ?ppc
Update recipe from 1.4.35 to 1.4.39.
Drop the second chunk of the patch for 1.4.35, as it is already in 1.4.39.
Regarding the first chunk, which replaces -std=gnu99 by -std=gnu9x in configure.ac, it seems we need to keep the change because gcc2 probably ignores that "-std=gnu9x was deprecated in favor of -std=gnu99" and, consequently, does not understand -std=gnu99.
Add ?x86 ?x86_64 ?arm ?ppc to ARCHITECTURES.
Switch from http to https for HOMEPAGE and SOURCE_URI as lighttpd.net is now using SSL certificates from https://letsencrypt.org/