lcab: fix patch to meet /bin/haikuporter requirements.
Although this recipe was OK for the haikuporter from the haikuporter repo, it was failling to pass the patch phase with the default haikuporter, i.e. the one that is available as a "pre-built" package from the haikuports repo and which gets installed in /bin.
This edit fixes that issue by switching the patch format from ".patch" to ".patchset" and, most important, avoiding the use of patches created with "diff -U0" since haikuporter does not accept contexts of 0 lines.