geoipupdate: create recipe and fix that of geoip.
Created recipe for version 2.2.2 with nice polishing for the man
pages, the runtime and the default configuration:
* Template config in system/data/geoipupdate/GeoIP.conf.default.
* Active config in system/settings/GeoIP.conf.
* Download location for all GeoIP Legacy and GeoIP2 databases:
  B_SYSTEM_CACHE_DIRECTORY/GeoIP/ (i.e. /boot/system/cache/GeoIP/).
* Patch for the default config to allow geoipupdate to download the
  GeoLite2-City and GeoLite2-Country databases out-of-the-box.

Updated the recipe for geoip-1.6.9 to make it use the same directory
used by geoipupdate instead of /var/lib/geoip/ previously used when
that recipe was created. Also fixed missing mkdir in geoipupdate.sh
and added lost patch with miscellaneous enhancements for it and some
cleaning in the recipe.