Marble: use a safe source.
The version we were using was not even a git tag in KDE git repo. I
tried to find a close enough (slightly more recent) tag, and pull it
from a github archive.

Change the version number to match the tag naming (1.9.2 used in the old
recipe is what it says in the About box, but that doesn't seem to change
as often as the tags).

Issue #1103