qemu: add recipe for version 2.9.0.
From the Changelog: WARNING! Haiku is marked as unsupported.
This release includes changes to the configure script so that it will
now report some host operating systems and platforms as "unsupported".
These are host setups which we do not have access to and are thus unable
to test. They will continue to work in this 2.9 release (though
configure will warn you about the unsupported status), but in a future
QEMU release we may drop support for those hosts unless somebody
volunteers to help us with maintaining them (and can provide build/CI
machines).