rust: add recipes for 1.9.0 and 1.10.0 (#1626)
The recipe for 1.9.0 will build the full compiler suite for x86_gcc2,
and it will 'wrap' a binary stage0 compiler for x86_64.
The recipe for 1.10.0 is a complete build.

These recipes don't create the rust_runtime packages. This package
should be recreated at a later point, but we will need to think about
the package naming, as it will be very much possible to have various
runtime libraries running alongside each other.