Rust: add recipe for rust 1.13.0
The difference with the recipes for versions 1.9.0 through 1.12.1 is that
this one is based off the "official" rust sources, after Alexander
von Gluck integrated the existing patches and added some additional of his
own. The recipe uses some additional patches to fix the build system (and
some oversights). The complete patched source tree can be found at
https://github.com/nielx/rust