Rust: several changes to the 1.15.1 recipe
Add the cargo command to the package. Rust's package management tool is now
included in the package. The rationale for inclusion versus external package
is that the cargo command is closely related to the version of the rustc
compiler, and with the future dependency of the rust build system on cargo
they are closely linked.