dev-libs/GeoIP: bump to 1.6.12, add debuginfo, shrink TEST(). (#2193)
Upstream ships a tiny GeoIP.dat since 1.6.10 so we can drop the
optional steps in TEST().

There seems to be a bug in curl (7.58.0 or earlier) which did not
exist in curl 7.50.3 when the patch for geoipupdate.sh was added.
This happens when the script tries to update an up-to-date GeoIP
database file. It looks like curl's "--time-cond" option forgets to
instruct "--remote-time" to skip its job if a "304 Not Modified" is
received. If the output file exists then this does not get noticed.
But if this file does not exist then curl complains with an unwanted
"Failed to set filetime ... on outfile" warning.
Trying to fix this issue in our script would require to redirect
curl's stderr to /dev/null (or creating a temporary (empty) file to
make curl happy) but doing so would not be a good idea.
We can live with this warning until curl gets fixed.