util-linux: bump to 2.32, build lib{blkid,fdisk,uuid}. (#2525)
Also build blkid, fdisk, hexdump and uuidgen, but put fdisk in a
subpackage, marked as untested, as the geometry detection is not
implemented for Haiku, so the default values are not useful at all.
fdisk is working fine but it does not know how to inform Haiku that
a partition table was modified. Workaround: reboot.