util-linux: fix INSTALL when ALLOW_UNTESTED=no
* While at it, do not export CFLAGS and LDFLAGS. Instead, pass
  CPPFLAGS and LDFLAGS to both autogen.sh and configure.
* Now that we don't override the project's default CFLAGS anymore
  we can add the debuginfo package.

Warning: If you call haikuporter after the recipe was last modified
and then change ALLOW_UNTESTED to yes, the .DependencyInfo file for
the fdisk subpackage will be missing. Workaround: touch the recipe
before calling haikuporter. Of course, this is required only once,
that is, if you have recently changed ALLOW_UNTESTED to yes.