Rust: Add recipe for 1.21.0
The rust_bin recipe is built of binary artifacts. The main reason is
that the 32 bit version can no longer be built on Haiku, because it
crosses a hard memory limit. Perhaps this changes later on. In any case,
an added advantage is that his makes the package buildable by the
buildbots.

For documentation purposes, a regular rust recipe is also provided. This
will probably run on a x86_64 machine with a decent amount of RAM. The
package depends on rustc and cargo being available for the same version.
This makes it both buildable from the rust_bin and the rust recipe.