ragel: Mark as broken on x86_gcc2, drop commandBinDir.
Unlike ragel 6.9, which builds on x86_gcc2 primary arch, 7.0.0.10
BUILD_REQUIRES devel:libcolm which is not available on x86_gcc2.
We therefore need to keep 6.9 for x86_gcc2.

So, in order to allow x86_gcc2 hybrid to be able to activate both
ragel-6.9 and ragel_x86-7.0.0.10, we cannot install cmd:ragel in
$commandBinDir; we need to install it in $binDir.

Minor change: drop the call to autoreconf as it is not needed.