Rust: Add recipe for 1.27.0
This change includes both the rust and rust_bin recipe. The remark regarding
the memory usage of the build on Haiku remains,that's why the build is marked
as untested. I have, however, tested the recipe up to the point where the
main build starts so the odds are good that it might build.

Also remove the rust~dev recipe for the 1.22 version.