gdb: Add almost-working recipe and patchset for 8.1 (!!!).
Reworking the Haiku backend from 6.3 (2004) to 8.1 (2018) surprisingly
only took a few hours to get it to compile ... but then it crashes when
attempting to start inferiors with a use-after free that I spent over
three hours debugging and got nowhere.

It seems that somehow, an event object that is created, deleted, and otherwise
managed entirely by the Haiku patches is somehow deleted between waits?!
The guarded heap was of not much help here, it obviously caught the UAF
before the regular heap did, but couldn't even say where it had been freed,
only where it had been allocated; and the allocation location looked
completely wrong also. So, this is very close, but it doesn't work just yet.

It would sure be nice to have Qt Creator's GDB support (which needs
Python-enabled GDB) working...