rhash: add commandSuffix & commandBinDir, rework package layout.
On all primary architectures, as well as on x86 secondary arch for
x86_gcc2h, the shared lib, the manual pages, the cmd:rhash runtime
and its symlinks all remain in the base package.

On x86 secondary arch for x86_64 hybrid the runtime is moved to
rhash_x86_bin, and the base package, rhash_x86, will only ship the
shared lib.