brotli: don't use $pythonVersion in REQUIRES_python* and BUILD_*.
The initial changes in brotli and pycurl to get rid of the "2.7"
and "3.6" were working on my system... but not on the buildbots.
Unfortunately, the fixes to get the recipes also work with the
buildbots have a small drawback: the dependencies evaluated at
build time are no longer tied with the minor version of Python.

So let's not replicate these changes to other recipes for now.