PLY: add python(2) subpackage, add TEST, add BUILD, fix HOMEPAGE.
Until now, INSTALL was doing both the build and the installation.
But this was bad because INSTALL was cleaning the build directory
of the python subpackage in order to build the python3 subpackage.
Creating separate build trees for each of the python subpackages
allows to keep both build trees. We can now use a TEST function.