golang: Update patchset to enable bootstrapping. (#3793)
This patches the go compiler for Haiku to fix process
waiting when executing Wait() on a process. Should also fix
the net/url hang when attempting to bootstrap Go 1.5 and
above.

Also changes the GOROOT_FINAL location to a globally writeable
location in the non-packaged folder to allow golang to install
third-party libraries.