icu: preliminary recipe for version 64.2.
There is no way this will build for gcc2. Versions of ICU from 59
onwards require and use C++11. This means anything wanting to use the
new APIs will not build with gcc2 anymore.

Marked as untested for now, still need to investigate:
- Why embedding the data in libicudata doesn't work, gcc fails to link
the lib, but manually re running the command works?
- Migration process to get Haiku to use it on all architectures, and
wether to keep an icu57 package around for some time.