SDLLopan: use a darker tileset
The tileset tile0.pcx had the tiles using a white background.
That made it impossible for the game to highlight a selected tile.
Edited tile0.pcx (using GIMP) to use a darker background.

Improved the SUMMARY - nobody cares if it uses SDL.
Improved the DESCRIPTION - since it's all part of an unchanging
package, no need to explain how to add tilesets/backgrounds. That's
up to the packager.