ICU: update to 57.2
Includes more up to date tzdata (2019a instead of 2016f in the previous
recipe), which should get us more in line with the latest timezone
reforms. Also includes support for the new Japanese era Reiwa (令和).

No API changes, so this should be a drop-in replacement for ICU 57.1.

We should investigate updating to newer versions for non-gcc2 platforms,
however (but this requires ABI changes so rebootstrapping is needed).