icu: recipe for version 66.1.
We should rebootstrap all the non-gcc2 architectures using this. New
versions of WebKit require at least ICU60 and we are currently shipping
version 57.

Currently the recipe has a different name so it can be installed
side-by-side with the old one, but we don't want to ship Haiku with 2
versions of ICU.

In order to not further delay the beta2 release, this should be done
after the release.