icu 57: disable devel package for non-gcc2
This will make sure any future builds will use icu 66 instead on
platforms where it is available. This should make it possible to
smoothly migrate to a newer ICU version without breaking old packages.

When we're sure nothing depends on icu 57 anymore we can see about
removing it (it will remain for gcc2 of course, since it's not possible
to build later versions there).