Noto: bump version
This version looks quite bad if you are using greyscale antialiasing +
hinting on small font sizes (the hinting is too strong). But it looks
better with the LCD subpixel mode (or at larger font sizes where the
hinting isn't that strong probably).

I guess this will help avoiding the consequences of #15007.