first attempt at splitting up the package
add a texmf and texmf_doc subpackage containing $dataDir/texlive/texmf-dist{,/doc}
this is both to work around Haiku ticket #12116 (and still have docs) as well as
making the package more managable in general.

Ideally the package should be split up according to texlive collections, but
that would require parsing of texlive.tlpdb of some sort.