Texlive: bump to 2021 (#5963)
* add texlive 2020

changes in addition to bump and required changes:
 - uses system libraries where feasible instead of bundled ones
 - uses install-tl and tlpdb from a source tarball instead of an SVN checkout (removes the dependency on svn)

* first attempt at splitting up the package

add a texmf and texmf_doc subpackage containing $dataDir/texlive/texmf-dist{,/doc}
this is both to work around Haiku ticket #12116 (and still have docs) as well as
making the package more managable in general.

Ideally the package should be split up according to texlive collections, but
that would require parsing of texlive.tlpdb of some sort.

* replace envsubst with sed replacements

envsubst doesn't work because haikuporter's variables are not exported and are only available in the recipe itself

* replace the tlpdb tarball with the extra tarball which contains some of the yet missing files

also, change the SELFAUTO patch dynamically in the recipe to use $dataDir (i.e. a /packages based path) instead of hard-coding /boot/system/data. This way the installed tools can also be used during the INSTALL phase in the haikuports recipe itself.

* files that are copied directly by install-tl need to be copied and not linked (otherwise it copies the symlink which doesn't work in the installed system)