texlive: split texlive_texmf into multiple subpackages (per collection)
There is one package for the runtime files, documentation and source files
for each TeXLive collection each. Packages that would be empty are omitted.
Additionaly, there are '_full' metapackages that depend on all the others.

The subpackage metadata and packageEntries are generated using perl scripts
that use the TeXLive package database API.

The package metadata is pre-generated because it needs already extracted and
preprocessed sources.