Blind fix for the radare2 pkg
The shlr lib is not suposed to be installed in the system, and the list after prepareInstalledDevelLibs was incomplete. i just updated this, is the CI going to verify that im not pushing something wrong? i dont have time for testing or learning how to build ports from haiku right now, so i would appreicate some help/feedback here if needed.

thanks