gcc: remove hack that breaks everything.
gcc pretended to be tinycc to workaround the fact that Haiku could not
handle a version of gcc larger than 9. However, this led to cmake
thinking that the compiler was tinycc, and disabled all optimizations in
release builds. Which in turn made webkit fail to build.

Everything that was compiled with gcc11 and cmake until now needs to be
rebuilt to have optimizations.