x11-libs: Mark libraries that actually utilize the X connection directly.
Add a comment and make them depend on libX11_xcb instead of just libX11.
The libraries which do not actually use the connection but just call APIs
are left alone.

Preparation for BeXlib (which is a ways off, but this will allow me
to build things in HaikuPorts against it properly.)