binutils: add recipe for version 2.36 and remove some old recipes
This version of binutils supports PE output, which is used for the 32bit
EFI bootloader. Fixes https://dev.haiku-os.org/ticket/17568.

Previous upgrades to binutils had been left disabled because of a bug
(linked in the recipes), but the bug has since been closed (not sure if
it was actually fixed). The only way to see if this will break something
is to try to update binutils again, and see if anything breaks.