hstr: fix configuration instructions and other minor changes. (#7270)
* Use the proper configuration instructions for bash in Haiku.
  Big thanks to madmax for finding the issue!
* Reworked the patch about the history files locations.
* Removed the TEST() section, as "make check" had "nothing to do".
* Removed "-I m4" from aclocal, as it was not necessary.
* Removed the "hh" symlink to "hstr". The user can create an alias.

The issue was that, on platforms that lack the TIOCSTI ioctl, a
workaround is needed to be able to insert text on the shell's command
line. The workaround relies on defining a shell function,
and keybinding said function so it gets handled by readline.

Now, following the "hstr --show-configuration" instructions will
result in C-r working as expected, at least for bash.

As it still won't fully work under zsh (for the lack of a similar
workaround, but implemented as a zsh function)... removed the ZSH
mention from the package description to avoid false expectations.