wireshark: separate summary/description for gui sub-package
It's a good idea to add separate summaries/descriptions for
sub-packages, where it makes sense.

Here, for example, there's the main standalone Wireshark
package, which is set to appear in the "Featured packages" while
the optional GUI is in a "non-featured" package.
By referencing the GUI package people have an easier time finding
out about it.

Not bumping revision to save people downloading 40MiB.