python3.10: bump version to 3.10.8. (#7313)
* python310: bump revision to 3.10.8.

* Use "make altinstall".
* Documented problematic tests, and skipped the ones that hang/stall.
* Enable optimizations.
* Added a note regarding tests that invoke the crash dialog.

Running "hp --test python310" now gives the same results[*] for both
non-optimized, and optimized builds, so let's just enable
optimizations for 3.10.8.

[*] As there's some randomness involved in the test runs, YMMV.

Anecdotal/aditional info:

Building and testing with "--disable-test-modules" might be
considered? Python's docs read:

   "The test package is meant for internal use by Python only."

And that those are used for regression testing mostly. See:
https://docs.python.org/3/library/test.html#module-test

Personal experience: I got more consistent results with that flag
enabled, than witout it. I left it out, trying to avoid hiding
problems (a.k.a.: failing silently. A big NO-NO in QA/QC), and
because I assume that packaging Python counts as "internal use".

* python310: patch tests that need several GB of storage.

They now behave like on Windows or MacOSX, requiring to be enabled
by the use of the "largefile" resource usage flag.

* python3.10: rename package from python310 to python3.10.

This matches the rename of python3 to python3.7, and the similar
change for Python 3.9.

Renamed patchset files to match the recipe naming.

Sorted sections according to the Haikuporter Guidelines.