zsh: fix prompt configuration
Create a default zshrc file that sets up the prompt to behave the same
as the bash one set by Haiku. There was an attempt to patch the prompt
in zprofile already (after sourcing /etc/profile), but the prompt is now
too complex for the simple substitutions to work.

Avoids the problem reported in https://dev.haiku-os.org/ticket/17652