links: fix build on x86 (32 bits).
This seems to have been broken since the update to 2.28, as 2.21
is the last version available on the repositories.

The fix avoids errors resulting from conflicting integer types
between the <tiff.h> and <SupportDefs.h> headers.

Needed to also specify "-g -O2 -fopenmp" when defining CFLAGS to
avoid linking errors (and to match the default flags used otherwise).

None of this is necessary on x86_64, thus the conditional on "_x86".