texlive: bump version to 2023
Building this requires ICU 70 or newer and building libxml2 (an indirect dependeny) against ICU 70
to get rid of icu66_devel in the build environment. These changes are done in separate commits.

ConTeXt is not included in TeXLive 2023. If it is needed, it should probably be built as a
separate package (the `mtxrun` call was commented out from texlive_update.sh before, so it
probably didn't work before anyway). See https://tug.org/texlive/bugs.html for some details.