jsonrpcclient: drop Python 3.7 support, add it for 3.9/3.10.
Also: Fixed REQUIRES. jsonrpclient still won't run as the version of
the `click` package in the repos (8.1.3) is too new for this version
of `jsonrpclient` (requires click < 7.x).

Updating to 4.0.3 will require either switching to install from
the wheel (.whl) file provided by the project over PyPi, or having
a working Poetry package in Haiku.

Packaging Poetry and its nearly 30 dependencies is still an on-going
quest, so in the meantime we might want to use that .whl after all.