lapack: switch to Python 3.9
Moved cmd:python3 to TEST_REQUIRES, as even if it seems like the
build needs it (for "summary tests"), it actually doesn't, and its
only for "make test" we need Python3.

Added $jobArgs, as that speeds up the build quite a lot.

Testes passed 100% with or without that set, in any case.