texlive: fix references to texlive_core package links directory
This would bind the texlive packages to a fixed version (and revision!) of texlive_core.
Also, they couldn't be shared between x86 and other architectures, because the packages are named
`texlive_core_x86` on x86 because of secondaryArch.

Revbump for rebuild (texlive was built against texlive_core 2022 before and is therefore broken right now.)