chocolate_doom: recipe cleanups.
* Drop usage of Python 2.x.
* Drop python_imaging dependency.
* Drop _x86 suffixes.
* Include basic docs.
* Create _debuginfo package.

Python dependency is optional, needed for generating man pages, for example.

Those weren't included before, and to switch to Python 3, we need to
update to newer versions of chocolate_doom (no new release > 3.0.1 yet),
so just drop that requirement.

"python_imaging" can be replaced by "pillow" (which supports python3),
but it is really only needed to re-create font files, or convert icons.

Neither is strictly needed when building under Haiku.

Instead if providing a _debuginfo package, we might want to just strip
the binaries ("make install-exec-strip").