yt-dlp: switch to nightly builds
The last 'official' release was in March 2023. Development of yt-dlp
is still very busy, but they switched to providing nightly builds [1],
while possibly also doing 'stable' releases from time to time in the
future.

I think we should switch to their nightly builds as well, as long as
they appear to be stable, and graba new one every month or so.

[1] https://github.com/yt-dlp/yt-dlp/issues/7175#issuecomment-1569442827