emacs: update to 29.1, add patch to disable `be-resources`
The `be-resources` command doesn't run properly inside the chroot and isn't
necessary since we use our own rdef to customize the application.

It's also possible to fix `be-resources` on x86_64 by adding
`haiku_datatranslators` to the BUILD_PREREQUIRES but this doesn't fix the 32
bit build.  The translation roster still doesn't work right.