contourpy: new recipe.
Listed as a dependency for matplotlib.

Only providing a Python 3.10 for now, as this is a new recipe, and its
only inmediate target user would be matplotlib (that should also be
moved to 3.10).

In any case, I've tested the recipe works for multiple versions,
by locally building it for both 3.9 and 3.10.

Note: built (and very lightly tested) on 64 bits only.