alien_sdl: switch to secondaryArch, add tests, cleanup
The patch function is replaced with a patchset. For proper secondaryArch support, more patching will likely be necessary.

TODO: additional_cflags in ConfigData are still wrong (they use the wrong include paths).
They should be fixed or removed if not needed.

TODO: determine if the 'build_*' entries in ConfigData are needed.
If not, this could maybe be changed to an 'any' arch package.
(By default, this calls the `sdl-config` script from SDL anyway, so it depends on the environment
whether secondaryArch will be used or not.)