Rust: cleanup recipes and WIP for 1.73.0
This commit cleans out all previous Rust recipes, and provides a work
in progress recipe for 1.73.0.

The current status of the 1.73.0 recipe is that it contains all changes
that are needed to do a successful build. I am running into an error,
where the intermediate stage1 compiler is not able to find the core
and standard libraries. This is not an issue building outside of
`haikuporter`. I suspect the chroot environment may be the reason, but
it requires more investigation.