fuse-nfs: add disabled recipe for 1.0.0~git (#9999)
Mounting and listing files/directires works for NFSv3/NFSv4.

Copying files from an NFSv3 share causes an SMAP violation and stalls with NFSv4.

It doesn't seem possible to unmount a share.  Tracker doesn't give the option in the context
menu and running `unmount` in a Terminal fails with a device busy message.