android_file_transfer: enable FUSE userlandfs add-on
I got this working some time ago but then I never got to enable it in the
recipe. It allows to access a mounted Android phone in Tracker, no Qt
needed.

Split the Qt app into a separate package so that doing this does not
require installing Qt. It allows to access a mounted Android phone in
Tracker, no Qt needed.

Split the Qt app into a separate package so that doing this does not
require installing Qt..