libsdl2: remove unneeded extra looper
When SDL2 is running in an application that already has a main thread,
it started its own looper to handle things. This create problems because
this thread can end up calling application code, and the code is
confused about which thread it is running in.

Instead, create a BHandler and attach it to the existing BApplication
looper, so all code is run from the same thread.