dma: add recipe for 0.14~git
Untested with internet delivery but delivering to local users works.

The recipe is a bit ugly since the `dma` executable wants to have setgid of the
group that is created by the package itself.  The package could be made to run
as uid 0, but, using a post-install script to make the spool dirs and a copy of
the dma executable with correct ownership seems to work ok for now.