fs_uae: Update to 3.1.66
In order to get --with-glew to work, I had to manually add all of the C
flags and linker options via env variables. glew.pc eventually tries to
include gl.pc which doesn't exist. However, it seems that I only need to
provide -Lgl and the program should work correctly (and it does).