fs_uae, fs_uae_launcher: Update FS-UAE to 3.1.66, patch Launcher for better exe finding (#10455)
* fs_uae_launcher: Better Haiku-friendly binary search

When updating fs-uae to 3.1.66, fs-uae-launcher was unable to find the
fs-uae binary when the launcher was launched from the Deskbar. This
changes the binary search code for Haiku to always look in the same
direcctory where fs-uae-launcher is to be installed.

* fs_uae: Update to 3.1.66

In order to get --with-glew to work, I had to manually add all of the C
flags and linker options via env variables. glew.pc eventually tries to
include gl.pc which doesn't exist. However, it seems that I only need to
provide -Lgl and the program should work correctly (and it does).

* Indent recpies using tabs