perl: bump version, enable threads, don't install sitecustomize.pl
See https://perldoc.perl.org/perl#History for details of changes.
Fixes CVE-2023-47038. (the other security fix affects only Windows)

Because this is a major update of perl, this requires rebuilding all perl packages.

Enable threads, which are enabled by most Linux distros as well, apparently.
This required some additional patches.

Don't install the `sitecustomize.pl` script any more. I don't think it is actually needed.
If it is retained though, it would need some changes to support secondaryArch.
The patches which create the script are still included for now, but they could be removed
if this change is accepted.
Fixes #10624.