python_pkgconfig: recipe cleanup/fixes.
While this module could be "any" in theory, doing so means the "x86_gcc2" version
of "pkg-config" gets called (that causes problems when packages that use this module
to build stuff, get the wrong info on 32 bits, no sane python package would be asking
for, or expecting, GCC2 stuff).

Cleaned up a bit the recipe while at it, and droped 3.9, as the only user on-tree
of this module is Python 3.10 only already (borgbackup).