paramiko: recipe cleanup. (#10765)
Just a quick cleanup. Moving it to "any", as this is just a pure-python module.

Tests do not run, as we're missing the "invoke" depedency for "pytest.relaxed",
and that means bringing in at least 6 more new recipes.

Similar with trying to update paramiko to newer versions.

Seems like too much work for a package that is not used by anything in-tree (thus,
just this clean up for now).