wxgtk: update to libtiff 6, disable fswatcher
wx detects that we have kqueue, and tries to use it for filesystem
notify operations. But we don't implement that part of kqueue currently.
For now, disable filesystem notify support, until either we get a better
kqueue implementation, or someone writes a backend based on the node
monitor for wx.