guile 3.0: new version of guile (#10798)
The library can be be installed separately from guile 2.2 (and guile 1).

The old version of the package is renamed to guile2.2, its devel and tools packages conflict with the new version. It can be removed as soon as all dependent packages are moved to 3.0.

Lilypond is still built with guile 2.2, as it is the recommended version for the current stable branch. The unstable version switched to guile 3.0 exclusively, so this is a preparation for that.