xml_sax: new perl module recipe
This includes a patch to install the ParserDetails.ini configuration file into
site_perl (i.e. the non-packaged hierarchy) instead of vendor_perl because it
be editable by other perl module packages that provide an XML SAX parser
(whether installed as HPKG or manually).

The config file is created and modified using post-install and pre-uninstall
scripts.